requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
alembic==1.18.3
    # via -r requirements.in
annotated-doc==0.0.4
    # via fastapi
annotated-types==0.7.0
    # via pydantic
anyio==4.12.0
    # via
    #   starlette
    #   watchfiles
bcrypt==3.2.0
    # via -r requirements.in
cffi==2.0.0
    # via
    #   bcrypt
    #   cryptography
click==8.3.1
    # via uvicorn
cryptography==46.0.3
    # via python-jose
dnspython==2.8.0
    # via email-validator
ecdsa==0.19.1
    # via python-jose
email-validator==2.3.0
    # via -r requirements.in
fastapi==0.127.1
    # via -r requirements.in
greenlet==3.3.0
    # via sqlalchemy
h11==0.16.0
    # via uvicorn
httptools==0.7.1
    # via uvicorn
idna==3.11
    # via
    #   anyio
    #   email-validator
mako==1.3.10
    # via alembic
markupsafe==3.0.3
    # via mako
paho-mqtt==2.1.0
    # via -r requirements.in
passlib==1.7.4
    # via -r requirements.in
psycopg2-binary==2.9.11
    # via -r requirements.in
pyasn1==0.6.2
    # via
    #   python-jose
    #   rsa
pycparser==2.23
    # via cffi
pydantic==2.12.5
    # via
    #   -r requirements.in
    #   fastapi
pydantic-core==2.41.5
    # via pydantic
python-dotenv==1.2.1
    # via
    #   -r requirements.in
    #   uvicorn
python-jose[cryptography]==3.5.0
    # via -r requirements.in
pyyaml==6.0.3
    # via uvicorn
rsa==4.9.1
    # via python-jose
six==1.17.0
    # via
    #   bcrypt
    #   ecdsa
sqlalchemy==2.0.45
    # via
    #   -r requirements.in
    #   alembic
starlette==0.50.0
    # via fastapi
typing-extensions==4.15.0
    # via
    #   alembic
    #   anyio
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   starlette
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
uvicorn[standard]==0.40.0
    # via -r requirements.in
uvloop==0.22.1
    # via uvicorn
watchfiles==1.1.1
    # via uvicorn
websockets==15.0.1
    # via uvicorn
